\name{discreteMTP-package}

\alias{discreteMTP}

\docType{package}

\title{
Multiple Testing Procedures For Discrete Test Statistics
}

\description{
Multiple testing procedures for discrete test statistics, that use the known discrete null distribution of the p-values for simultaneous inference. 
The package contains a \code{\link{p.discrete.adjust}} function,  similar to \code{\link{p.adjust}}, that provides the adjusted p-values from two step-up and two step-down multiple testing procedures that exploit the discrete null distribution of the p-values. 
A detailed description of these procedures can be found in Heller & Gur (2011) \href{http://arxiv.org/abs/1112.4627v1}{link}. 
}

\details{
\tabular{ll}{
Package: \tab discreteMTP\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2012-04-28\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Ruth Heller, Hadas Gur and Shay Yaacoby.

Maintainer: Shay Yaacoby \email{shay66@gmail.com}.
}
